flatten the first ring of the sphere model
When zoomed into ground level, the low resolution geometry
causes the curvature of the Earth to become visible when
it should appear flat.

There are several options to address this problem.

1. Increase the resolution of the sphere
2. Procedurally render the sphere
3. Artifically flatten the top ring of the sphere to make
   the ground appear flat when near ground level

Option 3 was the simplest to implement for the demo and
Trekking Maps already renders the Earth at high resolution
near ground level so this problem will not persist.